pyproject.toml
[tool.poetry]
name = "e-library"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "e_library"}]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
django = "^4.1.7"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
drf-spectacular = "^0.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
